dubbo增加group,增加zookeeper锁

+++ src/main/java/com/ykly/service/impl/GouldServiceImpl.java
@@ -122,7 +122,7 @@ public class GouldServiceImpl implements GouldService {
     
     @Override
     public String redisLock(String key, String value) {
-        long flag = redisUtil.setnx(key, value);
+        int flag = redisUtil.setnx(key, value);
         if(flag == 1){
             logger.info("redisUtil setnx:{}", key);
             redisUtil.del(key);
